backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncmy==0.2.11
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-docshield==0.2.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
limits==5.8.0
packaging==26.0
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scalar_fastapi==1.7.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
